Migrate monitoring station view page (#1340)
* Migrate monitoring station view page

https://eaflood.atlassian.net/browse/WATER-4679

As the legacy licence view page as now been rebuilt, the next process is to update some of the linked pages
to legacy pages. One of these legacy pages is the monitoring station view. If we are able to move this over to
our system, it will help towards removing the use of old legacy code.

This PR is focused on rebuilding the monitoring station view page from the legacy code base.