Al's take on Licence import flagging supp. billing
https://eaflood.atlassian.net/browse/WATER-4591

There are so many dates to contend with, and possible scenarios to cover that providing guidance over a flaky video call is proving difficult. Also, I've not been focused or directly hands on with this (till this branch!) so I haven't been providing the clearest guidance.

So, this is me putting my money where my mouth is, and building on what has been done to try and provide a working 'take' on how this might be done a little simpler and with less processing.