Housekeeping - rename service that does processing
Because we stole the name `Determine` for our other service, we needed to rename this one. The fact we have to pass in the 'imported' licence made me reconsider trying to avoid referring to this being import related.

I think it can stay here: it is about processing supplementary billing flags for a licence, so `licences/supplementary` is a better home IMHO.

But let's make it clear in the name this is the service that deals with imported licences.