Add a new fetch existing licence details service
Because of the complexity of the query being made (not 1, not 2, but 3 EXISTS calls!) I'm comfortable that we drop out of Objection.js and into Knex `raw()`. The pattern to make the call and retrieve a single result (because there will only be one) is the same used during our work on the import itself.