Refactor main processing service
Focused on two things

- Can we zero in on the earliest end date
- doing one 'thing' per step

Hopefully this makes sense. Essentially, grab the data we need to make decisions about the flags. Then use that data in functions dedicated to each type of flag.

I know that the values they return cannot be directly passed to the persist service, but it's a small step to tweak it.