Refactor logic for creating return cycles when importing new licence (#1389)
https://eaflood.atlassian.net/browse/WATER-4543

As part of the work to replace the legacy system NALD we need to generate return logs. This PR creates a return cycle if one does not already exist when creating return logs for a new licence when it is imported. This will create the return cycles even if the date for them to be ran has not yet happened.