Fix issue in view
The problem is caused because the description is being added to a non-markup object `<a>`. This means Nunjucks is auto-escaping the content to protect us from an XSS attack.

To tell it not to auto-escape we have to use the [safe filter](https://mozilla.github.io/nunjucks/templating.html#autoescaping).