Return log start and end dates to be determined using the correct cycle dates (#1397)
https://eaflood.atlassian.net/browse/WATER-4543

As part of the work to replace the legacy system NALD we need to generate return logs. This PR updates the logic around how the start, end and due dates are calculated based on the time of the year the request is being made. This is mostly for new licences that have been imported or for when licences are ended and will need new return logs created.