Handle missing Monitoring station metadata
https://eaflood.atlassian.net/browse/WATER-4679

We are working through the legacy pages and migrating them to this new repo. We've been working on migrating the view monitoring station page and another issue has been found during the latest round of QA.

For some records, we're showing the **WISKI ID** as 'null'. As a double check, we've looked at the data and found all three metadata properties (Grid reference, WISKI ID, and Station reference) can be either `NULL` or `''` in the DB. When it value is `''` it looks like we are handling it. But because we're actually not, when the value is `NULL` it comes through to the UI.

This change ensures we are properly handling these properties when they are 'blank' in the DB.