Add option to run tests without cleaning
https://github.com/DEFRA/water-abstraction-team/issues/124

Over the last few months we have been working through our test suite removing all calls to `DatabaseSupport.clean()`. Our aim was to get to a position that the database is only cleaned before the test suit runs.

We're there now, which means not only is our test suite faster, but we are able to consider moving to a modern test framework that will run tests in parallel.

To get to this position, not only do we clean the DB before each test run, we also seed all the reference data the tests need. We see the benefit when running the full test suite compared to how things were previously. But, when you are working on a feature and only wanting to run specific tests, this upfront step is very noticeable.

In theory, the changes we've made _should_ mean cleaning the DB is unnecessary. In fact, its a handy way of spotting tests that are likely to be 'flaky'! But for now, it is still prudent to have clean then test as the default. However, it would be handy to have the option to skip the clean in order to speed up our TDD red-green-refactor cycle.

This change adds a new `test:skip` script to `package.json`. The call is the same as the current one for `test`. The fact it isn't called `test:` means [npm](https://www.npmjs.com/) won't call `pretest:` before it is run. Simples!