Handle missing Monitoring station metadata (#1405)
https://eaflood.atlassian.net/browse/WATER-4679

We are working through the legacy pages and migrating them to this new repo. We've been working on migrating the view monitoring station page; another issue was found during the latest round of QA.

We show the **WISKI ID** as 'null' for some records. As a double check, we've looked at the data and found that all three metadata properties (Grid reference, WISKI ID, and Station reference) can be either `NULL` or `''` in the DB. When the value is `''`, it looks like we are handling it. But because we're actually not, when the value is `NULL`, it comes through to the UI.

This change ensures we correctly handle these properties when they are 'blank' in the DB.