Migrate licence document to have data fro unique constraint
https://eaflood.atlassian.net/browse/WATER-4708

As part of the ongoing work to move the import code into WRLS. We have needed to alter the database view in order to enforce the unique constraints on a table.

In this case we need to add defaulted data from document type and regime. This data will likely always be the default.

This change add these two columns to the view to enable an upsert to take place on the unique constraint.