Spike creating the 2PT supplementary bill run
https://eaflood.atlassian.net/browse/WATER-4201

User acceptance testing of the two-part tariff annual bill run is almost complete. There appear to be just a few niggles and 'nice to haves' remaining.

We've moved onto repurposing this work for supplementary and made changes to support this. For example, allowing users to select 'Two-part tariff supplementary' as an option when creating a bill run (see the epic WATER-4202 for all changes). Also, lots of work to support how we flag licences for inclusion in a two-part tariff supplementary bill run.

We're now at the point of needing to stitch it all together. After a user selects two-part tariff supplementary, select the licences to pass to the match and allocate engine. Then, after reviewing the matches and confirming the results, ensuring we create the actual supplementary bill run correctly.

This spike is a space to get the complete journey working, and work out what exactly is needed.