Import licence document for a licence (#1392)
* Import licence document for a licence

https://eaflood.atlassian.net/browse/WATER-4669

We need to replace the import service logic to import a licence from NALD.

WRLS has the concept of a licence document. NALD does not.

As part of the import process this licence document is created from the NALD licence. The start and end date use the licence and licence versions to calculate the earliest date for both.

This change will replicate this logic.

We will insert this imported data in the relevant public views.

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>