Fix auth scope on view monitoring station route (#1413)
https://eaflood.atlassian.net/browse/WATER-4679

We are working through the legacy pages and migrating them to this new repo. We've been working on migrating the view monitoring station page; another issue was found during the latest round of QA.

When we added the new route, we set the auth scope to `billing`. Clearly a copy and paste error, because when the legacy code is reviewed it has no scope.

Like viewing a licence, all users should be able to view a monitoring station. Once in the page their role will determine whether they can do anything else with it.

This change removes the scope to fix the issue.