Rename gauging stations to monitoring stations
https://eaflood.atlassian.net/browse/WATER-4679

When we first added gauging stations as a model to this project it was as part of our work to migrate the view licence page. On its summary tab it will display any gauging stations the licence is linked to.

Though that's not really true! What it displays are the 'Monitoring stations' a licence is linked to. If you click through one that is listed, the page you'll see will be on the url `/monitoring-stations` (both legacy and ours). If you use the main search the results are displayed under the section title 'Monitoring stations'.

When we did the migration of the view licence page, we did challenge why we still refer to them as 'Gauging stations' and whether we should be updating the UI to match.

The answer we got back was that though 'Gauging stations' is commonly used by the users we speak to, they are only one type of station that we have in our system. So, when this feature was added it was agreed to use the generic term 'Monitoring stations'.

What we now have is one of those discrepancies we often find in the legacy code which causes confusion. The DB and/or code is named differently to what we see in the UI.

So, before we do anything too adventurous with them, this change updates anything named 'gauging station' to be 'monitoring station'.