Remove POST create bill run route
https://eaflood.atlassian.net/browse/WATER-4201

When we first added SROC supplementary billing to this project, creation of the bill run was triggered from [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui).

Later, we've added SROC two-part tariff annual billing. That required a change to the create bill run journey, so we took the opportunity to migrate it to this project.

This means we no longer expect the legacy UI to trigger bill runs. The end point it was using is now defunct. But also, it means we can simplify our bill run creation/initialisation logic because we no longer have to consider there being two trigger points.

The thing we have been working on recently is SROC two-part tariff supplementary. We're ready to stitch all the pieces we've built together, the first step being triggering the bill run. Our journey now includes the option, but it hasn't been updated to check whether a user _can_ create a two-part tariff supplementary bill run (based on state of existing bill runs).

That's going to complete the create/initiate process even further. So, as a preparatory step, we're simplifying things first by dropping the old POST create bill run route.