2PT use full authorised volume rather than Nil (#1404)
https://eaflood.atlassian.net/browse/WATER-4684

From testing the business has requested that in circumstances where we can't match or determine a volume for the review stage, we use the Full Authorised Volume rather than a nil volume. Unless there are multiple returns and multiple charge versions where only some match, in which case use the volumes available.

This PR will resolve this issue.