Rename Rtn reqs. setup folder and route 'method'
https://github.com/DEFRA/water-abstraction-team/issues/126

We are on the cusp of our work to migrate the management of return versions (requirements) from NALD to WRLS going live.

This has been a major feature that has seen much iteration since we first started. So, before it gets used in earnest we want to ensure it is 'ship shape and Bristol fashion'!

A key bit of the tidy up is to remove all modules related to setting up a return version into `/setup` sub folders, using the same pattern we did for setting up a bill run.

The problem is, we already used 'setup' for one of the steps in the journey. 😩

So, this change renames the route `/method` as well as all related modules.