Import licence document roles for a licence - licence holder
https://eaflood.atlassian.net/browse/WATER-4717

We need to replace the import service logic to import a licence from NALD.

WRLS has the concept of a licence document. NALD does not.

As part of the import process a licence document is created from the NALD licence.

There are two roles we want to import returnsTo and licence holder.

This change will focus on the licence holder licence role.

We will insert this imported data in the relevant public views.

Previous work to set handle import was done for the returnTo role here - https://github.com/DEFRA/water-abstraction-system/pull/1414