Rename Rtn reqs. setup folder and route 'method' (#1428)
https://github.com/DEFRA/water-abstraction-team/issues/126

We are on the cusp of our work to migrate the management of return versions (requirements) from NALD to WRLS going 'live'.

This is a significant feature that has seen many iterations since we started. So, before it gets used in earnest, we want to ensure it is 'ship shape and Bristol fashion'!

A key part of the tidy-up is moving all modules related to setting up a return version into a `/setup` subfolder, using the same pattern we used for setting up a bill run.

The problem is that we already used 'setup' for one of the steps in the journey. 😩

So, this change renames the route `/method` and all related modules.