Licence import flagging for supplementary billing (#1338)
https://eaflood.atlassian.net/browse/WATER-4591

During the licence import process, we need to check if a licence has been lapsed/revoked or ended. A licence can end at any time, either in the past or the future. If the licence has ended in the past then the licence bills will be incorrect. This means the licence will need to be flagged for either pre-sroc supplementary billing, sroc supplementary billing or two-part tariff supplementary billing (depending on what is affected). This PR is building the service needed to check the licence lapsed/revoked or ended date and flagging if necessary.