Licence Import Supplementary billing test endpoint
https://eaflood.atlassian.net/browse/WATER-4591

Due to the fact the above ticket is happening during the licence import (which is automated), we need to add a test endpoint to allow the QA team to test the new supplementary flagging functionality.
The process of flagging a licence for supplementary billing only happens if there is a new end date added to the licence (lapsed, expired or revoked). The imported licence is looking at NALD data, so built into our testing endpoint is functionality that means you can pass a new licence end date and it will act as though that is the imported NALD data.