Refactor supplementary billing flag services
https://eaflood.atlassian.net/browse/WATER-4137
https://eaflood.atlassian.net/browse/WATER-4199
https://eaflood.atlassian.net/browse/WATER-4622
https://eaflood.atlassian.net/browse/WATER-4588

The work for flagging two-part tariff supplementary billing years has no be completed. There are multiple different ways these flags can be set, and each way had its own ticket. When starting the work for these tickets we came across some code that needed updated and some that wasn't working. This is specifically for flagging pre sroc and sroc supplementary as well. This meant we ended up migrating some legacy code to the system to flag for the pre sroc and sroc supplementary. Now all the work has been done, this PR is a refactor of the code, de-duping some of the functionality and tidying it up to make it more consistent.