Fix testing endpoint for supplementary flags
Recently a PR was raised to create a testing endpoint for the process of flagging a licence for supplementary billing during the import service. This looks at the licence end dates and works out what if any flags should be added to the licence. During testing the feature wasn't working as expected. After some investigation, it turns out that the testing endpoint wasn't confured correctly. When the arguments are passed in the body of the request it converts everything to String. This means the data we are mocking is in a String data type and not a Date data type, which it would be during the actual import. This PR fixed the object before being passed to the supplementary flagging process.