Refactor billing account contact details to model
https://eaflood.atlassian.net/browse/WATER-4057

We currently display a billing account's contact details when viewing a bill in a bill run. This is achieved by the `FetchBillingAccountService` and the `ViewBillPresenter`.

We have to show the _exact_ same details in the _exact_ same format as part of reviewing the return match and allocation results in a two-part part tariff bill run.

And we strongly suspect there will be more pages in the coming months that will also need to display this.

We've already started a [tidy up of our two-part tariff review code](https://github.com/DEFRA/water-abstraction-system/pull/1443). Moving both the fetch _and_ the presentation logic to somewhere reusable will help with that and put us in a better position for future changes.

Previously, when we have needed to fetch various pieces of data in order to make a determination about something on a model we've done it on the model directly using modifiers and instance methods.

This change refactors our existing billing account detail logic to follow this pattern.