Fix failing unit tests (#1445)
This change fixes some unit tests that were failing in CI.

Some changes to return cycles introduced an intermittent fail. We tracked it down to the order of the results causing an issue. If we set the order on the query we use we remove the intermittent fail.

But, because they were failing someone who should have known (👉 @Cruikshanks 😳 ) better slammed in a change and didn't spot that some of the fails were their own.

So, fixes for those failing tests are included in this change as well.