Fix env var for licence purposes view
https://eaflood.atlassian.net/browse/WATER-4705

As part of the work to move legacy pages into the new system, the licence purposes page was redeveloped and successfully merged in.
However upon testing, it was discovered that the link to the licence purposes page on the licence summary page was still redirecting to
the legacy page. Upon inspection, this was due to a misspelling of the environment variable feature flag that shows the new licence purposes page.

This PR fixes the misspelt environment variable and correctly redirects to the new licence purposes page.