Feature toggle the Legacy Process Licence service
https://eaflood.atlassian.net/browse/WATER-4728

As part of ongoing work to migrate the legacy import logic into this repo, we are adding a trigger to enable us to start using this new logic - https://github.com/DEFRA/water-abstraction-system/pull/1453

However, some of the logic needs to be turned off behind a feature flag so the other logic can go live.

This import process was initially designed to be kicked off by the existing import service. This change will remove the route and controller for this older idea.