Add `,` between abstraction points in point model (#1457)
* Add comma between abstraction-point in point model

https://eaflood.atlassian.net/browse/WATER-4705

During the testing of the licence purposes page, it was noticed that the legacy page included commas between
abstraction points. Within the new system, the abstraction points are generated via the `$describe` modifier on the
point model which is called when fetching licence purposes. To keep with the original legacy page format, we will
add commas to the `$describe` modifier for multiple abstraction points.

This PR will add commas between abstraction points created by the point model `$describe` modifier.