Decouple start date logic
https://eaflood.atlassian.net/browse/WATER-4687

As part of work to introduce quarterly returns we have noticed some logic leaking out of the start date functionality.

Outside of this service we do not care how the start date is calculated.

This change introduces a variable 'returnVersionStartDate' to the session for services outside of start to use.

Start date has been left as intact as possible and this change has been added when the start date saves the session so the internal workings have been left alone.