Create new base review presenter & add first helpler
Determining the charge periods for an element entails extracting a number of properties from various objects, and then passing the values through DetermineAbstractionPeriodService. The results then need to be formatted.

We've got 3 presenters that have to work this out, resulting in lots of duplication of the same code. To remove the duplication, we add a new `base-review.presenter.js`. We'll be adding some more duplicated presenter logic to it in the next few commits.