Fix active nav bar
Spotted when starting to fix the tests that where we are not specifying the active nav bar in the controller, we were returning 'search' instead of 'bill-runs'. This means the nav bar was flicking between options depending on what review page you were on.

To try and avoid this in future we set the active in _all_ controllers and avoid setting it in the services.