Add quarterly returns submissions to return versions
https://eaflood.atlassian.net/browse/WATER-4707

As part of the quarterly returns submissions work, we need to add a new checkbox to the additional-submission-options page for the Quarterly returns submissions.

This has introduced some logic to determine if the return version is for a water company and if the return version start date is beyond the start of quarterly returns.