Fix issue handling
Spotted that where there are no issues, our logic was written with the assumption the value we'd see in the DB was null. In fact, it is getting set to an empty string.

Really, it should be null. But fixing that feels outside the scope of this tidying up. So, instead we move to formatting of issues to a central place so we don't overlook this quirk in future.