Add fixture
We have a number of large data objects we need to generate in order to test the various presenters and services used by bill runs review.

Typical we declare these in the test files. But in this case we need to reuse the same data objects ion a number of places. So, we are taking advantage of the concept of a fixture file. We can declare these large objects in one place, and then grab them within the unit tests.

We do them as functions rather than static objects to avoid any issues with one test updating the object causing later tests to fail. Each time a test calls a fixture function, it'll get its own copy of the data object.