Add quarterly returns to return versions table
https://eaflood.atlassian.net/browse/WATER-4707

As part of the work to introduce quarterly returns in the system repo we need to add a column to mark a return version for quarterly returns.

The work to handle this logic is done in the system repo.

This change is just to introduce the quarterly returns in the previous water tables.