Update chg ref presenter to use new formatter
The reason we added the formatter was because the acceptance tests highlighted we were doing a mix of things depending on

- if the factors were set (not 1)
- if the factors were set but they had been amended to 1 (they've been unset)

Basically, they were hiding and displaying, and the logic was having to mix the original and amended values to control this behaviour.

We've now simplified it. If you can amend the factors, we _always_ display them. Plus, we show the amended and original so the two values can be easily compared.

This simplifies the logic, which is what allowed us to move a chunk of the 'adjustments' code into a shared formatter.